# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Maintenance Copilot",
    "description": "Multi-agent industrial assistant powered by LangGraph, SQLite checkpointing, FAISS RRF, and Argon2 RBAC.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/telemetry/events": {
      "post": {
        "tags": [
          "Telemetry v1"
        ],
        "summary": "Ingest Telemetry Event",
        "operationId": "ingest_telemetry_event_api_v1_telemetry_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEventPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Telemetry v1"
        ],
        "summary": "List Telemetry Events",
        "operationId": "list_telemetry_events_api_v1_telemetry_events_get",
        "parameters": [
          {
            "name": "machine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by machine identifier",
              "title": "Machine Id"
            },
            "description": "Filter by machine identifier"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum events to retrieve",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum events to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/samples": {
      "post": {
        "tags": [
          "Telemetry v1"
        ],
        "summary": "Ingest Telemetry Samples",
        "description": "Ingest continuous sensor readings (spindle speed, temperature, vibration, motor current).",
        "operationId": "ingest_telemetry_samples_api_v1_telemetry_samples_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetrySampleBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Telemetry v1"
        ],
        "summary": "List Telemetry Samples",
        "description": "Continuous sensor readings for one machine, oldest first, with each metric's unit.",
        "operationId": "list_telemetry_samples_api_v1_telemetry_samples_get",
        "parameters": [
          {
            "name": "machine_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Machine identifier",
              "title": "Machine Id"
            },
            "description": "Machine identifier"
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "spindle_speed",
                    "temperature",
                    "vibration_rms",
                    "motor_current"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO8601 window start",
              "title": "From"
            },
            "description": "ISO8601 window start"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryEventPayload": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id",
            "description": "Unique UUID for idempotency"
          },
          "machine_id": {
            "type": "string",
            "title": "Machine Id",
            "description": "Target machine identifier, e.g. EQ-1001, EQ-2001"
          },
          "metric": {
            "type": "string",
            "title": "Metric",
            "description": "Telemetry metric name, e.g. vibration_rms, plasma_rf_forward_power"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Numeric measurement value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Measurement unit, e.g. mm/s, W, Torr"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "ISO8601 timestamp"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity",
            "description": "Explicit severity override: normal, low, medium, high, critical"
          },
          "fault_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fault Code",
            "description": "Optional linked fault code"
          },
          "raw_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Payload",
            "description": "Arbitrary raw sensor context"
          }
        },
        "type": "object",
        "required": [
          "machine_id",
          "metric",
          "value",
          "unit"
        ],
        "title": "TelemetryEventPayload"
      },
      "TelemetrySample": {
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "spindle_speed",
              "temperature",
              "vibration_rms",
              "motor_current"
            ],
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "value",
          "observed_at"
        ],
        "title": "TelemetrySample"
      },
      "TelemetrySampleBatch": {
        "properties": {
          "machine_id": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Machine Id"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TelemetrySample"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "machine_id",
          "samples"
        ],
        "title": "TelemetrySampleBatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}